requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-crontab==0.7.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
